+++ Common/socketHelpers.c
@@ -23,27 +23,27 @@ int64_t bulk_read(int fd, char *buf, size_t count){
     int c;
     size_t len=0;
     do{
-        c=TEMP_FAILURE_RETRY(read(fd,buf,count));
+        c= (int) TEMP_FAILURE_RETRY(read(fd, buf, count));
         if(c<0) return c;
-        if(c==0) return len; //EOF
+        if(c==0) return (int64_t) len; //EOF
         buf+=c;
         len+=c;
         count-=c;
     }while(count>0);
-    return len ;
+    return (int64_t) len;
 }
 
 int64_t bulk_write(int fd, char *buf, size_t count){
     int c;
     size_t len=0;
     do{
-        c=TEMP_FAILURE_RETRY(write(fd,buf,count));
+        c= (int) TEMP_FAILURE_RETRY(write(fd, buf, count));
         if(c<0) return c;
         buf+=c;
         len+=c;
         count-=c;
     }while(count>0);
-    return len ;
+    return (int64_t) len;
 }
 
 int bind_local_socket(char *name){
